scraper/requirements.txt
aiohttp==3.11.11
aiosqlite==0.20.0
boto3==1.35.92
browserforge==1.2.1
fake_useragent==2.0.3
google_api_python_client==2.157.0
google_auth_oauthlib==1.2.1
httpx==0.28.1
imap_tools==1.8.0
loguru==0.7.3
numpy==2.2.2
pandas==2.2.3
protobuf==5.29.3
pyotp==2.9.0
python-dotenv==1.0.1
redis==5.2.1
Requests==2.32.3
scipy==1.15.1
thefuzz==0.22.1
tqdm==4.67.1
typing_extensions==4.12.2
zendriver==0.2.3
